Updated aix_lvol.py to support shrink=true/false
I needed aix_lvol.py to support the shrink option as used by lvol.py and merged these into aix_lvol.py. 
If the shrink option is not passed the default is yes/true. 
If (like in my case) you would like to specify that an LV should be created if not present and the default size should be X at creation time - but (crucially) that if it exists, skip the shrink part and continue then specify shrink=no/false and it will contine with: msg="Logical volume %s size is already %sMB or higher." 
This solves my specific situation, and I feel that this is a safe option to be added for aix_lvol.py - and crucially keeps it logically in line with lvol.py